+++ hass-addon-mbusd/run.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bashio
+
+DEVICE=$(bashio::config 'DEVICE')
+BAUDRATE=$(bashio::config 'BAUDRATE')
+MODE=$(bashio::config 'MODE')
+LOGLEVEL=$(bashio::config 'LOGLEVEL')
+
+bashio::log.info "Starting mbusd -d -L - -v $LOGLEVEL -p $DEVICE -s $BAUDRATE -m $MODE -P 502"
+
+exec /usr/src/mbusd -d -L - -v $LOGLEVEL -p $DEVICE -s $BAUDRATE -m $MODE -P 502