+++ hass-addon-sunsynk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## **2021.9.20-0.0.6** - 2021-09-30
+
+- Allow TCP connections (PORT_ADDRESS)
+- Sunsynk: Async connect
+- Sunsynk: Faults & connection exceptions
+
 ## **2021.9.19-0.0.5** - 2021-09-22
 
 - Updated definitions & decoding